+++ R/df1.R
@@ -1,0 +1,12 @@
+#' Example dataset with longitudinal counts
+#'
+#' An example dataset of a study with longitudinal counts, used
+#' to illustrate how `ptmixed()` works.
+#' 
+#' @author Mirko Signorelli
+#' @docType data
+#' @keywords datasets
+#' @name df1
+#' @usage data(df1)
+#' @format A data frame with 18 rows and 5 variables
+NULL