+++ R/loglik.R
@@ -11,7 +11,7 @@
 #' @param Z Design matrix for the random effects
 #' @param id Id indicator (it should be numeric)
 #' @param offset Offset term to be added to the linear predictor
-#' @param GHk Number of quadrature points (default is 10)
+#' @param GHk Number of quadrature points (default is 5)
 #' @param tol Tolerance value for the evaluation of the probability mass function of the Poisson-Tweedie distribution
 #' @param GHs Quadrature points at which to evaluate the loglikelihood. If \code{NULL} (default value), 
 #' the GH quadrature points are computed internally
@@ -21,8 +21,8 @@
 #' @author Mirko Signorelli
 #' @seealso \code{\link{ptmixed}} and the examples therein
 
-loglik.pt.1re <- function(beta, D, a, Sigma, y, X, Z, id, offset = NULL, 
-                          GHk = 10, tol = 1e-323, GHs = NULL) {
+loglik.pt.1re = function(beta, D, a, Sigma, y, X, Z, id, offset = NULL, 
+                          GHk = 5, tol = 1e-323, GHs = NULL) {
   requireNamespace('tweeDEseq')
   requireNamespace('mvtnorm')
   with.offset = !is.null(offset)
@@ -31,15 +31,17 @@ loglik.pt.1re <- function(beta, D, a, Sigma, y, X, Z, id, offset = NULL,
   if (t %% 1 !=0) stop('The dataset appears to be unbalanced. The code for
                        the unbalanced case is not yet implemented (check back soon!)')
   if (is.null(GHs)) {
-    GHs <- GHpoints.pt.1re(y=y, id=id, X=X, Z=Z, beta=beta, D=D, a=a, 
+    GHs = GHpoints.pt.1re(y=y, id=id, X=X, Z=Z, beta=beta, D=D, a=a, 
                        Sigma=Sigma, offset=offset, RE.size = RE.size,
                        GHk=GHk, tol = tol)
   }
-  if (with.offset) Delta.ij <- c(X %*% beta + c(offset))
-  else Delta.ij <- c(X %*% beta)
+  if (with.offset) Delta.ij = c(X %*% beta + c(offset))
+  else Delta.ij = c(X %*% beta)
   # calculate log Lik
-  log.p.b <- t(sapply(GHs$scaled.b, mvtnorm::dmvnorm, mean = rep(0, RE.size), 
-                      sigma = Sigma, log = TRUE))
+  log.p.b = sapply(GHs$scaled.b, mvtnorm::dmvnorm, mean = rep(0, RE.size), 
+                      sigma = Sigma, log = TRUE)
+  if (GHk > 1) log.p.b = t(log.p.b)
+  else if (GHk == 1) log.p.b = as.matrix(log.p.b)
   aux = dim(GHs$scaled.b[[1]])[1]
   n = length(unique(id))
   mus = matrix(NA, n*t, aux) 
@@ -55,12 +57,12 @@ loglik.pt.1re <- function(beta, D, a, Sigma, y, X, Z, id, offset = NULL,
   }
   mus = exp(mus)
   #
-  ff <- function (yy) {
-    log.p.y.b <- rowsum(dpt.yvec.mumatr(yvec = yy, mu.matr = mus, D, a, 
+  ff = function (yy) {
+    log.p.y.b = rowsum(dpt.yvec.mumatr(yvec = yy, mu.matr = mus, D, a, 
                                         log = T, tol = tol), id)
     c((GHs$det.Bs * exp(log.p.y.b + log.p.b)) %*% GHs$wGH)
   }
-  p.y <- ff(y)
+  p.y = ff(y)
   p.y[which(p.y < 1e-323)] = 1e-323 # otherwise it's too small for log(p.y)
   #
   sum(log(p.y), na.rm = TRUE)